// packages/microservice/src/transporters/grpc/enums/kind.enum.ts
export enum Kind {
  Double = "double",
  Float = "float",
  Int32 = "int32",
  Int64 = "int64",
  Uint32 = "uint32",
  Uint64 = "uint64",
  Sint32 = "sint32",
  Sint64 = "sint64",
  Fixed32 = "fixed32",
  Fixed64 = "fixed64",
  Sfixed32 = "sfixed32",
  Sfixed64 = "sfixed64",
  Bool = "bool",
  String = "string",
  Bytes = "bytes",
}